Separate advanced search container on modules/components
