+++ client/src/containers/AdvancedSearch/IssueTable/logic/state.ts
@@ -1,0 +1,14 @@
+export interface FilterPartState {
+	id: string;
+	filterDef: WebApi.Entities.FilterDefinition;
+	searchText: string;
+	members: string[];
+}
+
+export interface AdvancedSearch {
+	filterParts: FilterPartState[];
+}
+
+export const initialState: AdvancedSearch = {
+	filterParts: [],
+};
